dashboard stat card

+++ modules/main/features/dashboard/case_daily_active_stats/repository/repository.go
@@ -107,7 +107,7 @@ WITH current_period AS (
   WHERE
     stat_date >= $1
     AND stat_date <  $2
-    AND ($3 IS NULL OR department_id = $3)
+    AND ($3::INT IS NULL OR department_id = $3::INT)
 ),
 previous_period AS (
   SELECT
@@ -116,7 +116,7 @@ previous_period AS (
   WHERE
     stat_date >= $4
     AND stat_date <  $5
-    AND ($3 IS NULL OR department_id = $3)
+    AND ($3::INT IS NULL OR department_id = $3::INT)
 )
 SELECT
   c.value AS value,
