+++ modules/main/features/dashboard/case_daily_remake_stats/repository/repository.go
@@ -148,7 +148,7 @@ WITH current_period AS (
   WHERE
     stat_date >= $1
     AND stat_date <  $2
-    AND ($3 IS NULL OR department_id = $3)
+    AND ($3::INT IS NULL OR department_id = $3::INT)
 ),
 previous_period AS (
   SELECT
@@ -157,7 +157,7 @@ previous_period AS (
   WHERE
     stat_date >= $4
     AND stat_date <  $5
-    AND ($3 IS NULL OR department_id = $3)
+    AND ($3::INT IS NULL OR department_id = $3::INT)
 )
 SELECT
   COALESCE(c.rate, 0)             AS rate,
